llm_analytics/requirements.txt
# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0

# Data
pandas>=2.0

# LLM Providers (install the one you need)
groq>=0.4.0
# openai>=1.0
# anthropic>=0.30.0

# Prompts
pyyaml>=6.0
